Enforce correct settings for each strategy
* Strategy initialization now checks that the settings provided to the
constructor are the correct type specified by the _settings_cls class
attribute of any Strategy subclass. This has the effect that a
Strategy author must explicitly specify what settings they are will to
accept. This deviates slightly from the gufe Protocol class, which
doesn't perform such a check. Tests were added to
`test_strategy_abstraction.py` showing only specified StrategySettings
will be allowed.

* Added TypeVar for ProtocolResults in the `Strategy` base class. Any
downstream use of the `propose` method will then check that it's being
fed a dictionary with subclasses of ProtocolResults. I don't believe
there is an easier way to do this since the ProtocolResult is embedded
in a dictionary.

* gufe doesn't work with pydantic v2, so I now use the pydantic v1
api.

* The match statement for capturing the max_runs / cutoff settings was
updated and properly catches which of these were set to None