Increase testing coverage
* Provide a variety of parameters to ConnectivityStrategySettings

Termination logic within _propose depends on the combination of
parameters given to ConnectivityStrategySettings. A preset batch of
valid settings are provided through SETTINGS_VALID.

* Test ConnectivityStrategySettings validators

Add a test for sets of ConnectivityStrategySettings including both
valid and invalid settings. If an Exception is expected, check that it
is raised, otherwise check that the settings could be instantiated.

* Simulate the ConnectivityStrategy

Test that the ConnectivityStrategy terminates after a set number of
iterations. The max number of iterations is set to 100.

* Remove abstract method bodies from coverage reports