Add gufe submodule for testing until gufev1.2.0 release
Since features in the planned gufev1.2.0 release are needed by
stratocaster, we need to accurately recreate an environment that gufe
will provide upon installation. By using a submodule, we can install
the base testing environment from the gufe environment.yml file. This
will be reverted once gufev1.2.0 is available on conda-forge.

This also makes it easier to react to last minute changes to gufe
prior to its v1.2.0 release.